forse funziona bene

+++ Printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mfanelli <mfanelli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 13:36:11 by mfanelli          #+#    #+#             */
-/*   Updated: 2024/12/02 14:27:54 by mfanelli         ###   ########.fr       */
+/*   Updated: 2024/12/02 16:12:03 by mfanelli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ int	ft_putnbrhex_min(int nbr);
 int	ft_putnbrhex_max(int nbr);
 int	ft_putnbrhex_ptr(unsigned long nbr);
 int	ft_putptr(unsigned long long ptr);
-int	ft_unsigned_putnbr(unsigned int n);
+int	ft_unsigned_putnbr(unsigned long n);
 int	ft_check(va_list args, const char *format, int i, int len);
 
 #endif
